Start Scene (Host & Client Button)
I've installed a plugin to make development easier, which allows us to see PlayerPrefs values on an editor window.

I'm using PlayerPrefs to store the user-selected option (host or client) and pass it to the gameplay scene, so that the appropriate function can be called there depending on whether the user chose to be a host or a client.